requirements.txt
Flask
requests
python-dotenv
flask-login
psycopg2-binary
black 
flask-sqlalchemy